Add files via upload
Incluido o autocomplete e preço atual do fundo, na tabela

+++ eventoapp01/src/main/java/com/entra21/eventoapp/models/TabelaFii.java
@@ -72,7 +72,7 @@ public class TabelaFii implements Serializable {
 		return PRECO_ATUAL;
 	}
 
-	public void setPRECO_ATUAL(float PRECO_ATUAL) {
+	public void setPRECO_ATUAL(double PRECO_ATUAL) {
 		this.PRECO_ATUAL = PRECO_ATUAL;
 	}
